fraud-detection-ai/data-pipeline/requirements.txt
# Data Pipeline Dependencies
kafka-python==2.0.2
redis==4.5.4
pydantic==1.10.7
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
lightgbm==3.3.5

# Stream Processing
apache-flink==1.17.0  # Alternative: use pyflink

# Geospatial calculations
geopy==2.3.0
geographiclib==2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging and monitoring
structlog==23.1.0
prometheus-client==0.16.0

# Development and testing
pytest==7.3.1
pytest-asyncio==0.21.0
black==23.3.0
isort==5.12.0
mypy==1.3.0
